Change the method for calling xfuzz in tests.
Modify all tests so that they now start a new process with the command

        python3 -m xfuzz ...

so that it's no longer necessary to call the fuzz() function directly.
In addition to giving some extra implementation flexibility, this also
makes it a little bit easier to capture the output of the fuzzer and
use it for additional tests.