Various improvements to the test suite
Modify the test suite so that the only tests that pytest now runs by
default are the ones in test/test_cli.py -- this will help to avoid some
confusion when students see some of the tests being skipped. It's still
possible to run the full test suite with `pytest test/`, and you can run
the full test suite minus the CLI tests with

        pytest test/ --skip-cli-tests

In addition, this update removes some of the additional test flags that
were accumulating to enable / disable certain test cases, and separates
out the "extra" tests into test/test_extra.py.